Device Files Updated
This is still a work in progress. This is the first of many changes needing to happen. These changes are moving the device files more in line with the Virgin Mobile Evo V instead of the Sprint Evo. These files still need lots of work at all the mk files need to be updated to point to all new files I have added. All the newly added files are from the stock ROM that shipped with the phone.